more refactoring

+++ scripts/network-override.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 # SETTINGS
-dns="127.0.0.1"
+ENABLE_TRAFFIC_SPLIT=true
+ENABLE_DNS_SPLIT=true
+
 DNS_RESOLVE="127.0.0.1"
 DNS="1.1.1.1 1.0.0.1"
 GATEWAY="A.A.A.A" # CHANGE THIS! TODO: detect automatically
@@ -13,23 +15,30 @@ if [[ $EUID -ne 0 ]]; then
 fi
 
 # VPN Traffic split
-# If your utun interface tend to change like mine you can do the following given the right interface is the last one in the list
-INTERFACE=$(ifconfig | grep -A1 utun | grep -B1 'inet ' | grep utun | cut -d : -f 1 | grep utun | tail -n 1)
+if [ "$ENABLE_TRAFFIC_SPLIT" = true ]; then
 
-# last line in the file is the gateway ip
-route change default ${GATEWAY}
-for NET in ${VPN_NETWORKS}; do route add -net ${NET} -interface $INTERFACE; done;
+    # If your utun interface tend to change like mine you can do the following given the right interface is the last one in the list
+    INTERFACE=$(ifconfig | grep -A1 utun | grep -B1 'inet ' | grep utun | cut -d : -f 1 | grep utun | tail -n 1)
+
+    # last line in the file is the gateway ip
+    route change default ${GATEWAY}
+    for NET in ${VPN_NETWORKS}; do
+        route add -net ${NET} -interface $INTERFACE
+    done
+fi
 
 # DNS
-# Update DNS in resolv.conf
-cat <<EOF | sudo tee /etc/resolv.conf
-nameserver ${DNS_RESOLVE}
-EOF
+if [ "$ENABLE_DNS_SPLIT" = true ]; then
 
-# Update DNS in system db
-STATES=$(scutil <<< list | awk '{ print $4 }' | egrep '^State:/Network/Service/[^/]+/DNS$' | grep -v 'gpd.pan')
-while read STATE; do printf "d.init\nget ${STATE}\nd.add ServerAddresses * ${DNS}\nset ${STATE}\nquit\n" | sudo scutil; done <<<"$STATES"
+    # Update DNS in resolv.conf
+    sudo echo "nameserver ${DNS_RESOLVE}" > /etc/resolv.conf
 
-# Flush
-sudo killall -HUP mDNSResponder
-sudo dscacheutil -flushcache
+    # Update DNS in system db
+    for STATE in $(scutil <<< 'list ^State:/Network/Service/[^/]+/DNS$' | awk '{ print $4 }' | grep -v 'gpd.pan'); do
+        printf "d.init\nd.add ServerAddresses * ${DNS}\nset ${STATE}\nquit\n" | sudo scutil
+    done
+
+    # Flush
+    sudo killall -HUP mDNSResponder
+    sudo dscacheutil -flushcache
+fi
